refactor: create view and controller in the MVC pattern
