+++ src/main/java/racingcar/view/InputView.java
@@ -1,0 +1,22 @@
+package racingcar.view;
+
+import camp.nextstep.edu.missionutils.Console;
+import java.util.Arrays;
+import java.util.List;
+
+public class InputView {
+
+    public List<String> getCarNames() {
+        System.out.println("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)");
+        List<String> names = Arrays.asList(Console.readLine()
+                .replace(" ", "")
+                .split(",", -1));
+
+        return names;
+    }
+
+    public int getTotalRounds() {
+        System.out.println("시도할 횟수는 몇 회인가요?");
+        return Integer.parseInt(Console.readLine());
+    }
+}
